% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis.R
\name{read_txt}
\alias{read_txt}
\title{Read a text file as a character vector}
\usage{
read_txt(file)
}
\arguments{
\item{file}{A single .txt file}
}
\value{
A character vector
}
\description{
Read a text file as a character vector
}
\examples{
read_txt("text.txt")
}
