% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgoR-package.R
\docType{package}
\name{upgoR-package}
\alias{upgoR}
\alias{upgoR-package}
\title{upgoR: Comparing Your Text Against Common Words}
\description{
This package helps you to craft clear, simple text but comparing your existing text against a list of the most commonly-used words in the English language.
}
\author{
\strong{Maintainer}: Lee Durbin \email{l.d.durbin@gmail.com}

}
\keyword{internal}
